requirements.txt
torch==1.3.1
numpy==1.18.1
matplotlib==3.1.3
gym==0.16.0
PyYAML==5.3.1
